# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BusinessNow PSA Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /phases:
    get:
      operationId: listPhases
      tags: [phases]
      summary: List phases for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of phases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phase"
    post:
      operationId: createPhase
      tags: [phases]
      summary: Create a phase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhaseBody"
      responses:
        "201":
          description: Created phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
  /phases/{id}:
    get:
      operationId: getPhase
      tags: [phases]
      summary: Get phase by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
        "404":
          description: Not found
    patch:
      operationId: updatePhase
      tags: [phases]
      summary: Update phase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhaseBody"
      responses:
        "200":
          description: Updated phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
    delete:
      operationId: deletePhase
      tags: [phases]
      summary: Delete phase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── TIMESHEETS ────────────────────────────────────────────────
...
  /invoices/{id}/line-items:
    get:
      operationId: listInvoiceLineItems
      tags: [invoices]
      summary: List line items for an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of line items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceLineItem"
    post:
      operationId: createInvoiceLineItem
      tags: [invoices]
      summary: Add a line item to an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceLineItemBody"
      responses:
        "201":
          description: Created line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceLineItem"
  /invoices/{id}/line-items/{lineItemId}:
    patch:
      operationId: updateInvoiceLineItem
      tags: [invoices]
      summary: Update a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lineItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceLineItemBody"
      responses:
        "200":
          description: Updated line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceLineItem"
    delete:
      operationId: deleteInvoiceLineItem
      tags: [invoices]
      summary: Delete a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: lineItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/line-items/autofill:
    post:
      operationId: autofillInvoiceLineItems
      tags: [invoices]
      summary: Auto-fill line items from approved time entries (T&M)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated line items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceLineItem"

  # ─── RESOURCE REQUESTS ─────────────────────────────────────────
...
  /opportunities:
    get:
      summary: List opportunities
      operationId: listOpportunities
      parameters:
        - in: query
          name: accountId
          schema:
            type: integer
        - in: query
          name: stage
          schema:
            type: string
        - in: query
          name: ownerId
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Opportunity'
    post:
      summary: Create opportunity
      operationId: createOpportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOpportunityBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
  /opportunities/{id}:
    get:
      summary: Get opportunity
      operationId: getOpportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
    patch:
      summary: Update opportunity
      operationId: updateOpportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOpportunityBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Opportunity'
    delete:
      summary: Delete opportunity
      operationId: deleteOpportunity
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
  /opportunities/{id}/convert-to-project:
    post:
      summary: Create a project from a Won opportunity
      operationId: convertOpportunityToProject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertOpportunityBody'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConvertOpportunityResult'
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        name:
          type: string
        status:
          type: string
        ownerId:
          type: integer
        startDate:
          type: string
        dueDate:
          type: string
        billingType:
          type: string
        budget:
          type: number
        trackedHours:
          type: number
        allocatedHours:
          type: number
        budgetedHours:
          type: number
        completion:
          type: integer
        health:
          type: string
        description:
          type: ["string", "null"]
        internalExternal:
          type: string
        isAdminProject:
          type: integer
        opportunityId:
          type: integer
          nullable: true
        deletedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - accountId
        - name
        - status
        - ownerId
        - startDate
        - dueDate
        - billingType
        - budget
        - trackedHours
        - allocatedHours
        - budgetedHours
        - completion
        - health
        - internalExternal
        - createdAt
        - updatedAt
...
    Phase:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        status:
          type: string
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        order:
          type: integer
        isSharedWithClient:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - projectId
        - name
        - status
        - order
        - isSharedWithClient
        - createdAt
    CreatePhaseBody:
      type: object
      properties:
        projectId:
          type: integer
        name:
          type: string
        status:
          type: string
        startDate:
          type: string
        dueDate:
          type: string
        order:
          type: integer
        isSharedWithClient:
          type: boolean
      required:
        - projectId
        - name
    UpdatePhaseBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        startDate:
          type: string
        dueDate:
          type: string
        order:
          type: integer
        isSharedWithClient:
          type: boolean

    # ─── TIMESHEET ──────────────────────────────────────────────
...
    InvoiceLineItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitRate:
          type: number
        amount:
          type: number
        taxAmount:
          type: number
        billable:
          type: boolean
        timeEntryId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        role:
          type: ["string", "null"]
        order:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - invoiceId
        - description
        - quantity
        - unitRate
        - amount
        - taxAmount
        - billable
        - order
        - createdAt
    CreateInvoiceLineItemBody:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitRate:
          type: number
        amount:
          type: number
        taxAmount:
          type: number
        billable:
          type: boolean
        timeEntryId:
          type: integer
        userId:
          type: integer
        role:
          type: string
        order:
          type: integer
      required:
        - description
        - quantity
        - unitRate
        - amount
    UpdateInvoiceLineItemBody:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitRate:
          type: number
        amount:
          type: number
        taxAmount:
          type: number
        billable:
          type: boolean
        order:
          type: integer

    # ─── PHASE 3 SCHEMAS ───────────────────────────────────────────
...
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        name:
          type: string
        stage:
          type: string
        probability:
          type: integer
        value:
          type: number
        description:
          type: string
          nullable: true
        closeDate:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        projectId:
          type: integer
          nullable: true
        accountName:
          type: string
          nullable: true
        ownerName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - accountId
        - name
        - stage
        - probability
        - value
        - createdAt
        - updatedAt
    CreateOpportunityBody:
      type: object
      properties:
        accountId:
          type: integer
        name:
          type: string
        stage:
          type: string
        probability:
          type: integer
        value:
          type: number
        description:
          type: string
          nullable: true
        closeDate:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
      required:
        - accountId
        - name
    UpdateOpportunityBody:
      type: object
      properties:
        accountId:
          type: integer
        name:
          type: string
        stage:
          type: string
        probability:
          type: integer
        value:
          type: number
        description:
          type: string
          nullable: true
        closeDate:
          type: string
          nullable: true
        ownerId:
          type: integer
          nullable: true
        projectId:
          type: integer
          nullable: true
    ConvertOpportunityBody:
      type: object
      properties:
        name:
          type: string
        billingType:
          type: string
        startDate:
          type: string
        dueDate:
          type: string
        ownerId:
          type: integer
      required:
        - name
        - startDate
        - dueDate
    ConvertOpportunityResult:
      type: object
      properties:
        project:
          $ref: '#/components/schemas/Project'
        opportunityId:
          type: integer
      required:
        - project
        - opportunityId
...